Serial Number reponse decoded
The serial numbers were coming back from scan() as an array of array of numbers.
Taking each element in the large array at a time, then converting each element in
those smaller arrays to hex, then joining them together in reverse sequence tended
to give me the right serial. For some reason it always ends in 28 for the 2 I have
tested.
Next step is to try and poll them specifically for temperature.